+++ editor/editor-core/src/plugins/collab-edit/ui/to-avatar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AvatarProps } from '@atlaskit/avatar-group';
-import memoizeOne, { EqualityFn } from 'memoize-one';
+import memoizeOne from 'memoize-one';
 import { CollabParticipant } from '../types';
 import { ColoredAvatarItem } from './colored-avatar-item';
 
@@ -16,12 +16,8 @@ const toAvatar = (participant: CollabParticipant): AvatarProps => ({
   ),
 });
 
-const participantEquals: EqualityFn = ([aRaw], [bRaw]) => {
-  const a = aRaw as CollabParticipant;
-  const b = bRaw as CollabParticipant;
+export default memoizeOne(toAvatar, function participantEquals([a], [b]) {
   return (
     a.name === b.name && a.avatar === b.avatar && a.sessionId === b.sessionId
   );
-};
-
-export default memoizeOne(toAvatar, participantEquals);
+});