+++ design-system/code/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/code
 
+## 14.1.5
+
+### Patch Changes
+
+- [`b85e7ce12cd`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b85e7ce12cd) - Internal upgrade of memoize-one to 6.0.0
+
 ## 14.1.4
 
 ### Patch Changes