+++ design-system/onboarding/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/onboarding
 
+## 10.2.6
+
+### Patch Changes
+
+- [`b85e7ce12cd`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b85e7ce12cd) - Internal upgrade of memoize-one to 6.0.0
+
 ## 10.2.5
 
 ### Patch Changes