+++ elements/util-data-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/util-data-test
 
+## 16.0.1
+
+### Patch Changes
+
+- [`b85e7ce12cd`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b85e7ce12cd) - Internal upgrade of memoize-one to 6.0.0
+
 ## 16.0.0
 
 ### Major Changes