+++ jira/color-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/color-picker
 
+## 1.2.6
+
+### Patch Changes
+
+- [`b85e7ce12cd`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b85e7ce12cd) - Internal upgrade of memoize-one to 6.0.0
+
 ## 1.2.5
 
 ### Patch Changes