+++ media/media-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-table
 
+## 9.0.3
+
+### Patch Changes
+
+- [`b85e7ce12cd`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b85e7ce12cd) - Internal upgrade of memoize-one to 6.0.0
+
 ## 9.0.2
 
 ### Patch Changes